requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
python-dotenv==1.0.0
openai==0.28.1
google-api-python-client==2.108.0
markdown2==2.4.10
Pillow==10.1.0
